move proxy blueprint from yafowil.common to yafowil.proxy
